MAINT: Avoid overflow in comparison with int64 on shorter platforms
Maybe not the best way, but its functional and easy to grok and I
doubt it matters speed wise.

+++ numpy/random/_generator.pyx
@@ -4229,7 +4229,7 @@ cdef class Generator:
             elif colors.size > 0 and not np.issubdtype(colors.dtype,
                                                        np.integer):
                 invalid_colors = True
-            elif np.any((colors < 0) | (colors > INT64_MAX)):
+            elif np.any((colors < 0) | (colors > np.int64(INT64_MAX))):
                 invalid_colors = True
         except ValueError:
             invalid_colors = True
